fix: separate auth id
